pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "ffmpeg>=1.4",
    "jinja2>=3.1.6",
    "numpy>=2.2.5",
    "openai>=1.78.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "soundfile>=0.13.1",
    "uvicorn>=0.34.2",
]
